# api/openapi.yaml
openapi: "3.0.3"
info:
  title: PopcornVault API
  description: IPTV source and channel management API.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/health:
    get:
      operationId: healthCheck
      summary: Health check
      tags: [Health]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /api/sources:
    get:
      operationId: listSources
      summary: List all sources
      tags: [Sources]
      responses:
        "200":
          description: Array of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Source"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: addSource
      summary: Add a new source and trigger ingest
      tags: [Sources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSourceRequest"
      responses:
        "201":
          description: Source created and channels ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddSourceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/sources/{id}:
    parameters:
      - $ref: "#/components/parameters/SourceID"

    get:
      operationId: getSource
      summary: Get a source by ID
      tags: [Sources]
      responses:
        "200":
          description: Source detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: updateSource
      summary: Update mutable fields of a source
      tags: [Sources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSourceRequest"
      responses:
        "200":
          description: Updated source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteSource
      summary: Delete a source and cascade to its channels and groups
      tags: [Sources]
      responses:
        "204":
          description: Source deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/sources/{id}/refresh:
    parameters:
      - $ref: "#/components/parameters/SourceID"

    post:
      operationId: refreshSource
      summary: Re-ingest a source from its URL
      tags: [Sources]
      responses:
        "200":
          description: Source refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/channels/search:
    get:
      operationId: searchChannels
      summary: Semantic search for channels using AI embeddings
      description: >
        Search for channels using natural language queries. Requires VOYAGE_API_KEY to be configured.
        Returns channels ranked by cosine similarity to the query embedding.
      tags: [Channels]
      parameters:
        - name: q
          in: query
          required: true
          description: Natural language search query
          schema:
            type: string
        - name: source_id
          in: query
          description: Filter by source ID
          schema:
            type: integer
            format: int64
        - name: group_id
          in: query
          description: Filter by group ID
          schema:
            type: integer
            format: int64
        - name: media_type
          in: query
          description: "Filter by media type (0 = Livestream, 1 = Movie, 2 = Serie)"
          schema:
            type: integer
            enum: [0, 1, 2]
        - name: favorite
          in: query
          description: Filter by favorite status (true or false)
          schema:
            type: boolean
        - name: limit
          in: query
          description: "Max results to return (default: 20, max: 200)"
          schema:
            type: integer
            default: 20
            maximum: 200
      responses:
        "200":
          description: Channels ranked by semantic similarity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemanticSearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          description: Semantic search not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/channels:
    get:
      operationId: listChannels
      summary: List channels with optional filters and pagination
      tags: [Channels]
      parameters:
        - name: search
          in: query
          description: Case-insensitive substring match on channel name
          schema:
            type: string
        - name: source_id
          in: query
          description: Filter by source ID
          schema:
            type: integer
            format: int64
        - name: group_id
          in: query
          description: Filter by group ID
          schema:
            type: integer
            format: int64
        - name: media_type
          in: query
          description: "Filter by media type (0 = Livestream, 1 = Movie, 2 = Serie)"
          schema:
            type: integer
            enum: [0, 1, 2]
        - name: favorite
          in: query
          description: Filter by favorite status (true or false)
          schema:
            type: boolean
        - name: limit
          in: query
          description: "Max items to return (default: 50, max: 200)"
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          description: Number of items to skip
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated channel list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/channels/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Channel ID
        schema:
          type: integer
          format: int64

    get:
      operationId: getChannel
      summary: Get a single channel by ID
      tags: [Channels]
      responses:
        "200":
          description: Channel detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/channels/{id}/favorite:
    parameters:
      - name: id
        in: path
        required: true
        description: Channel ID
        schema:
          type: integer
          format: int64

    patch:
      operationId: toggleChannelFavorite
      summary: Set or unset a channel as favorite
      tags: [Channels]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleFavoriteRequest"
      responses:
        "200":
          description: Favorite status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  channel_id:
                    type: integer
                    format: int64
                  favorite:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/groups:
    get:
      operationId: listGroups
      summary: List groups, optionally filtered by source
      tags: [Groups]
      parameters:
        - name: source_id
          in: query
          description: Filter by source ID
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Array of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  parameters:
    SourceID:
      name: id
      in: path
      required: true
      description: Source ID
      schema:
        type: integer
        format: int64

  schemas:
    Source:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        url:
          type: string
        source_type:
          type: integer
          description: "0 = M3U, 1 = M3U Link, 2 = Xtream, 3 = Custom"
        use_tvg_id:
          type: boolean
          nullable: true
        user_agent:
          type: string
        enabled:
          type: boolean
        last_updated:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true

    Channel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        url:
          type: string
        group:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        media_type:
          type: integer
          description: "0 = Livestream, 1 = Movie, 2 = Serie"
        source_id:
          type: integer
          format: int64
        group_id:
          type: integer
          format: int64
          nullable: true
        favorite:
          type: boolean
        group_name:
          type: string
          nullable: true

    Group:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        image:
          type: string
          nullable: true
        source_id:
          type: integer
          format: int64

    APIError:
      type: object
      required: [status, error]
      properties:
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: HTTP status text
        detail:
          type: string
          description: Human-readable error detail

    AddSourceRequest:
      type: object
      required: [url]
      properties:
        name:
          type: string
          description: "Optional source name (defaults to \"m3u\")"
        url:
          type: string
          description: M3U URL to fetch and ingest

    AddSourceResponse:
      type: object
      properties:
        source_id:
          type: integer
          format: int64
        channel_count:
          type: integer

    UpdateSourceRequest:
      type: object
      description: All fields are optional; only provided fields are updated.
      properties:
        name:
          type: string
        url:
          type: string
        user_agent:
          type: string
        enabled:
          type: boolean

    ToggleFavoriteRequest:
      type: object
      required: [favorite]
      properties:
        favorite:
          type: boolean

    RefreshResponse:
      type: object
      properties:
        source_id:
          type: integer
          format: int64
        channel_count:
          type: integer
        refreshed:
          type: boolean

    ChannelListResponse:
      type: object
      properties:
        channels:
          type: array
          items:
            $ref: "#/components/schemas/Channel"
        total:
          type: integer
          description: Total matching channels (before pagination)
        limit:
          type: integer
        offset:
          type: integer

    SemanticSearchResponse:
      type: object
      properties:
        channels:
          type: array
          items:
            $ref: "#/components/schemas/SemanticResult"
        limit:
          type: integer

    SemanticResult:
      type: object
      properties:
        channel:
          $ref: "#/components/schemas/Channel"
        similarity:
          type: number
          format: double
          description: "Cosine similarity score (0 to 1, higher is more similar)"

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"